refactor kotlin to use proper cli framework
